+++ CHANGELOG.md
@@ -11,10 +11,16 @@
 
 ### Upgrading
 
-Using the Elixir [EventStore](https://github.com/slashdotdash/eventstore) with Commanded? Please ensure you upgrade the following dependencies:
+Please ensure you upgrade the following event store dependencies.
+
+Using the Elixir [EventStore](https://github.com/slashdotdash/eventstore):
 
 - `eventstore` to [v0.11.0](https://hex.pm/packages/eventstore)
 - `commanded_eventstore_adapter` to [v0.2.0](https://hex.pm/packages/commanded_eventstore_adapter)
+
+Using Greg's [Event Store](https://eventstore.org/):
+
+- `commanded_extreme_adapter` to [v0.3.0](https://hex.pm/packages/commanded_extreme_adapter)
 
 ## v0.13.0